crypto-volatility-index/test/adapter.test.ts
import { assert } from 'chai'
import { assertSuccess } from '@chainlink/external-adapter'
import { AdapterRequest } from '@chainlink/types'
import { getConfig } from '../src/config'
import { execute } from '../adapter'

describe('execute', () => {
  const jobID = '1'

  context('successful calls @integration', () => {
    const requests = [
      {
        name: 'id not supplied',
        testData: { data: {} },
      },
    ]

    requests.forEach((req) => {
      it(`${req.name}`, async () => {
        const data = await execute(req.testData as AdapterRequest, getConfig())
        assertSuccess({ expected: 200, actual: data.statusCode }, data, jobID)
        assert.isAbove(data.result, 0)
        assert.isAbove(data.data.result, 0)
      })
    })
  })
})
